* kill.cc (main): Chack pid against INT_MIN and (U)INT_MAX since pid_t 	is 4 byte on all platforms. 	* mkgroup.c (main): Check domlist[i].id_offset against UINT_MAX since 	DWORD is 4 byte on all platforms. 	* mkpasswd.c (main): Ditto.

+++ winsup/utils/kill.cc
@@ -260,13 +260,13 @@ out:
       if (!pid)
 	pid = strtoll (*argv, &p, 10);
       if (*p != '\0'
-	  || (!force && (pid < LONG_MIN || pid > LONG_MAX))
-	  || (force && (pid <= 0 || pid > ULONG_MAX)))
+	  || (!force && (pid < INT_MIN || pid > INT_MAX))
+	  || (force && (pid <= 0 || pid > UINT_MAX)))
 	{
 	  fprintf (stderr, "%s: illegal pid: %s\n", prog_name, *argv);
 	  ret = 1;
 	}
-      else if (pid <= LONG_MAX && kill ((pid_t) pid, sig) == 0)
+      else if (pid <= INT_MAX && kill ((pid_t) pid, sig) == 0)
 	{
 	  if (force)
 	    forcekill ((pid_t) pid, sig, 1);
